Complete voting process.
Vote grid is shown when a poll is open or user has voted.
If the grid is shown and the poll is closed, the grid is disabled.
Results are shown after the poll has closed or user has voted (although
backend doesn't currently return the results before the poll is closed).

- Sort polls in current polls list.
- Add possibility to populate VoteGrid with a Ballot map. The ballot
submitted by user is returned in result request.
- Add disabled property to VoteGrid*.
- Add list of polls and poll closing function to admin page.
- Add better poll result display.
- Add poll status text below poll heading.
- Replace poll id with poll name in heading if available.
- Ignore a11y check on control that is interactive for mouse
accessibility.